Fix: move `BigBioConfig` to `utils/configs.py` (#290)
* feat(utils): add BuilderConfig for BigBio

* feat(datasets): replace BigBioConfig class in datasets with utils/configs.py version

* feat(datasets): replace BigBioConfig class in template.py with utils/configs.py version

* fix: remove unnecessary import
