+++ biodatasets/nlmchem/nlmchem.py
@@ -14,13 +14,13 @@
 
 import os
 import re
-from dataclasses import dataclass
 from typing import Dict, Iterator, List, Tuple
 
 import bioc
 import datasets
 
 from utils import schemas
+from utils.configs import BigBioConfig
 from utils.constants import Tasks
 
 
@@ -57,17 +57,6 @@ _URLs = {
 _SUPPORTED_TASKS = [Tasks.NAMED_ENTITY_RECOGNITION, Tasks.NAMED_ENTITY_DISAMBIGUATION]
 _SOURCE_VERSION = "1.0.0"
 _BIGBIO_VERSION = "1.0.0"
-
-
-@dataclass
-class BigBioConfig(datasets.BuilderConfig):
-    """BuilderConfig for BigBio."""
-
-    name: str = None
-    version: str = None
-    description: str = None
-    schema: str = None
-    subset_id: str = None
 
 
 class NLMChemDataset(datasets.GeneratorBasedBuilder):
