+++ biodatasets/scitail/scitail.py
@@ -24,12 +24,12 @@ entails label and 16,925 examples with neutral label.
 """
 
 import os
-from dataclasses import dataclass
 
 import datasets
 import pandas as pd
 
 from utils import schemas
+from utils.configs import BigBioConfig
 from utils.constants import Tasks
 
 
@@ -67,17 +67,6 @@ _SUPPORTED_TASKS = [Tasks.TEXTUAL_ENTAILMENT]
 _SOURCE_VERSION = "1.1.0"
 
 _BIGBIO_VERSION = "1.0.0"
-
-
-@dataclass
-class BigBioConfig(datasets.BuilderConfig):
-    """BuilderConfig for BigBio."""
-
-    name: str = None
-    version: str = None
-    description: str = None
-    schema: str = None
-    subset_id: str = None
 
 
 class SciTail(datasets.GeneratorBasedBuilder):
