+++ examples/bioasq.py
@@ -28,8 +28,8 @@ import os
 import re
 import glob
 import datasets
-from dataclasses import dataclass
 
+from utils.configs import BigBioConfig
 from utils.constants import Tasks
 
 _CITATION = """\
@@ -374,16 +374,6 @@ _URLs = {
 _SUPPORTED_TASKS = [Tasks.QUESTION_ANSWERING]
 _SOURCE_VERSION = "1.0.0"
 _BIGBIO_VERSION = "1.0.0"
-
-
-@dataclass
-class BigBioConfig(datasets.BuilderConfig):
-    """BuilderConfig for BigBio."""
-    name: str = None
-    version: str = None
-    description: str = None
-    schema: str = None
-    subset_id: str = None
 
 
 class BioasqDataset(datasets.GeneratorBasedBuilder):
