+++ utils/configs.py
@@ -1,0 +1,13 @@
+import datasets
+from dataclasses import dataclass
+
+
+@dataclass
+class BigBioConfig(datasets.BuilderConfig):
+    """BuilderConfig for BigBio."""
+
+    name: str = None
+    version: datasets.Version = None
+    description: str = None
+    schema: str = None
+    subset_id: str = None
